Added a logout button to the SelectionAppBar and CharacterAppBar which allows the user to logout of their account additionally I moved out all the logic for the login from App.jsx to LoginSheet.jsx because firstly it is cleaner and secondly the state variables left on the App.jsx were being preserved throughout the app which were causing problems on logout
